requirements.txt
# Khai báo thư viện cần thiết
python-dotenv
smtplib
email